+++ src/container.lua
@@ -20,6 +20,8 @@ this program.  If not, see <https://www.gnu.org/licenses/>.
 
 require "src.utilities"
 
+global<const> *
+
 ---@build block:
 -- ------------------------------------------------------------------------- --
 --
@@ -31,7 +33,7 @@ require "src.utilities"
 ---@param container table
 ---@param pod table
 ---@param simulate boolean
-function container__create(container, pod, simulate)
+global function container__create(container, pod, simulate)
     -- main command
     local commands = { "podman run" }
 
@@ -110,7 +112,7 @@ end
 ---@param pod_name string
 ---@param container_alternate_name string
 ---@return boolean
-function container__ensure_name(container, pod_name, container_alternate_name)
+global function container__ensure_name(container, pod_name, container_alternate_name)
     -- container name is optional
     if string.is_nil_or_empty(container.name) then
         container.name = pod_name .. "-" .. container_alternate_name
@@ -127,7 +129,7 @@ end
 ---@param container table
 ---@param pod_name string
 ---@return boolean
-function container__is_valid(container, pod_name)
+global function container__is_valid(container, pod_name)
     if table.is_nil_or_empty(container) then
         log.error("A container in pod '" .. pod_name .. "' is empty!")
         return false
@@ -143,7 +145,7 @@ end
 ---Stop and removes a container.
 ---@param container table
 ---@param simulate boolean
-function container__remove(container, simulate)
+global function container__remove(container, simulate)
     system.exec("podman stop " .. string.escape_shell(container.name), "Stop container '" .. container.name .. "': ", simulate, false)
     system.exec("podman rm " .. string.escape_shell(container.name), "Remove container '" .. container.name .. "': ", simulate, false)
 end
@@ -152,7 +154,7 @@ end
 ---@param container table
 ---@param pod table
 ---@param simulate boolean
-function container__update(container, pod, simulate)
+global function container__update(container, pod, simulate)
     local registry = table.get_or_default(container, "registry", pod.registry)
     log.print("Update container '" .. container.name .. "' ...")
     system.exec("podman pull " .. string.escape_shell(registry .. "/" .. container.image), "", simulate, false)